refactor : 도메인 패키지에 user 서브 패키지 추가
